move and rename methods
